Convert WebpackService to TypeScript
Clear out the barrel so we don't get errors from the .js files.
