+++ package.json
@@ -29,6 +29,7 @@
     "@types/storybook__addon-actions": "^3.4.2",
     "@types/storybook__react": "^4.0.1",
     "@types/uuid": "^3.4.4",
+    "@types/webpack": "^4.4.26",
     "awesome-typescript-loader": "^5.2.1",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
