--- packages/brookjs-cli/src/services/index.js
@@ -1,7 +1,0 @@
-import * as hd from './hd';
-import * as template from './template';
-
-export { default as ui } from './ui';
-export { default as glob } from './glob';
-export { default as WebpackService } from './webpack';
-export { hd, template };
