+++ packages/brookjs-cli/src/services/index.ts
@@ -1,0 +1,1 @@
+export { default as WebpackService } from './webpack';
